requirements.txt
great_expectations==0.14.9
pyspark==3.2.1
pandas==1.4.1
tabulate==0.8.9
allure-pytest==2.9.45